+++ pytorch_lightning/core/lightning.py
@@ -32,7 +32,15 @@ else:
     XLA_AVAILABLE = True
 
 
-class LightningModule(ABC, DeviceDtypeModuleMixin, GradInformation, ModelIO, ModelHooks, DistributedStateHooks, Module):
+class LightningModule(
+    ABC,
+    DeviceDtypeModuleMixin,
+    GradInformation,
+    ModelIO,
+    ModelHooks,
+    DistributedStateHooks,
+    Module
+):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)